ai_service/requirements.txt
numpy
pillow
torch
torchvision
git+https://github.com/openai/CLIP.git
fastapi
uvicorn
joblib
scikit-learn==1.6.1
